+++ FSbrowser/FSbrowser.php
@@ -26,7 +26,7 @@ require_once(dirname(__FILE__).DIRECTORY_SEPARATOR.'extra-library.php');
 session_start();
 
 class FSbrowser{
-	public function Version($f=FALSE){ return '0.3.0'; }
+	public function Version($f=FALSE){ return '0.3.2'; }
 	public function Product_url($u=FALSE){ return ($u === TRUE ? "https://github.com/sentfanwyaerda/FSnode" : "http://sent.wyaerda.org/FSbrowser/".'?version='.self::Version(TRUE).'&license='.str_replace(' ', '+', self::License()) );}
 	public function Product($full=FALSE){ return "FSnode Browser".($full ? " ".self::version(TRUE).(class_exists('FSnode') && method_exists('FSnode', 'Product') ? '/'.FSnode::Product(TRUE) : NULL).(class_exists('Xnode') && method_exists('Xnode', 'Product') ? '/'.Xnode::Product(TRUE) : NULL) : NULL); }
 	public function License($with_link=FALSE){ return ($with_link ? '<a href="'.self::License_url().'">' : NULL).'cc-by-nd 3.0'.($with_link ? '</a>' : NULL); }
@@ -77,7 +77,7 @@ class FSbrowser{
 		if(is_object($this->handler)){
 			$lines = array();
 			/*fix*/ $subURI = str_replace(DIRECTORY_SEPARATOR, '/',  preg_replace("#^".$this->handler->realpath('/')."#", '', $this->handler->realpath($subURI)) );
-			/*fix*/ if(substr($subURI, 0, 1) != '/'){ $subURI = '/'.$subURI; }
+			/*fix*/ if(substr($subURI, 0, 1) != DIRECTORY_SEPARATOR){ $subURI = DIRECTORY_SEPARATOR.$subURI; }
 			/*debug*/ print '<!-- FSbrowser::build.$subURI: '.$subURI.' ['.$this->handler->is_dir($subURI).'] -->'."\n";
 			if($this->handler->is_dir($subURI)){
 				$list = $this->handler->scandir(str_replace(array('[',']','{','}'), array('\[','\]','\{','\}'), $subURI));
@@ -85,8 +85,8 @@ class FSbrowser{
 				foreach($list as $f){
 					if($this->handler->realpath($subURI. DIRECTORY_SEPARATOR .$f) && !($f == '.')){
 					$lines[$f] = parse_template('templates/line.html', array(
-						'file:name.full' => ($f == '..' ? 'Parent Directory' : $f.($this->handler->is_dir($subURI. DIRECTORY_SEPARATOR .$f) ? '/' : NULL)),
-						'file:url' => str_replace(DIRECTORY_SEPARATOR, '/',  './FSbrowser.php?URI='.($f == '..' ? DIRECTORY_SEPARATOR .$this->label.dirname($subURI): DIRECTORY_SEPARATOR .$this->label.$subURI. DIRECTORY_SEPARATOR .$f) ),
+						'file:name.full' => ($f == '..' ? 'Parent Directory' : $f.($this->handler->is_dir($subURI. DIRECTORY_SEPARATOR .$f) ? DIRECTORY_SEPARATOR : NULL)),
+						'file:url' => preg_replace('#[/'.DIRECTORY_SEPARATOR.']+#', '/', './FSbrowser.php?URI='.($f == '..' ? DIRECTORY_SEPARATOR .$this->label.dirname($subURI).DIRECTORY_SEPARATOR: DIRECTORY_SEPARATOR .$this->label.$subURI. DIRECTORY_SEPARATOR .$f.($this->handler->is_dir($subURI. DIRECTORY_SEPARATOR .$f) ? DIRECTORY_SEPARATOR : NULL)) ),
 						'file:type.class' => ($f == '..' ? 'parent-directory' : ($this->handler->is_dir($subURI. DIRECTORY_SEPARATOR .$f) ? 'directory' : strtolower(preg_replace("#^(.*)[.]([a-z0-9]+)$#i", "\\2", $f)).' '.preg_replace("#[^a-z0-9-]#i", "-", $this->handler->mime_content_type($subURI. DIRECTORY_SEPARATOR .$f)) )),
 						'file:modified.last' => date('d-M-Y H:i', $this->handler->filemtime($subURI. DIRECTORY_SEPARATOR .$f) ),
 						'file:size' => ($this->handler->file_exists($subURI. DIRECTORY_SEPARATOR .$f) && !$this->handler->is_dir($subURI. DIRECTORY_SEPARATOR .$f) ? $this->_humanizeFileSize( $this->handler->filesize($subURI. DIRECTORY_SEPARATOR .$f)) : NULL),
@@ -98,7 +98,7 @@ class FSbrowser{
 				ksort($lines);
 			}
 			
-			$set = array('title'=>'Index of '.$subURI.'/','connection:status'=>($this->handler->is_connected() ? 'connected' : 'unconnected'),'include:lines()'=>implode($lines),'footer'=>NULL,'URI'=>$this->URI,'label'=>$this->label,'meta-information'=>$this->build_URI_info($subURI));
+			$set = array('title'=>'Index of '.preg_replace('#[/]+#', '/', $this->handler->relativepath($subURI).($this->handler->is_dir($subURI) ? DIRECTORY_SEPARATOR : NULL)),'connection:status'=>($this->handler->is_connected() ? 'connected' : 'unconnected'),'include:lines()'=>implode($lines),'footer'=>NULL,'URI'=>$this->URI,'label'=>$this->label,'meta-information'=>$this->build_URI_info($subURI));
 			return parse_template('templates/overview.html', $set);
 		}
 		else{
@@ -110,16 +110,21 @@ class FSbrowser{
 		return parse_template('templates/connect.html', $set);
 	}
 	function build_URI_info($subURI=NULL){
+		$fullURI = $subURI.($this->handler->is_dir($subURI) ? '/' : NULL);
 		$set = array(
 		#	'' => $this->handler->($URI),
 			'subURI' => $subURI,
-			'URI' => $this->URI
+			'URI' => $this->URI,
+		#	'URI*' => $this->handler->parse_url($this->URI),
+			'realpath' => $this->handler->realpath($fullURI),
+			'realpath_URI' => $this->handler->realpath_URI($fullURI),
+			'relativepath' => $this->handler->relativepath($fullURI),
 		);
 		foreach(array('fileatime'=>2,'filectime'=>2,'filegroup'=>1,'fileinode'=>1,'filemtime'=>2,'fileowner'=>1,'fileperms'=>1,'filesize'=>3,'filetype'=>1,'file_exists'=>9,'is_dir'=>9,'is_executable'=>9,'is_file'=>9,'is_readable'=>9,'is_writeable'=>9,'disk_free_space'=>3,'disk_total_space'=>3,'mime_content_type'=>1 /*,'stat'=>5*/) as $method=>$action){
 			$set[$method] = $this->handler->$method($subURI);
 			switch($action){
-				case 2: /*timestamp*/ $set[$method] = date("Y-m-d H:i:s", $set[$method]); break;
-				case 3: /*filesize*/ $set[$method] = self::_humanizeFileSize($set[$method]); break;
+				case 2: /*timestamp*/ $set[$method] = ($set[$method] ? date("Y-m-d H:i:s", $set[$method]) : NULL); break;
+				case 3: /*filesize*/ $set[$method] = ($set[$method] >= 0 ? self::_humanizeFileSize($set[$method]) : NULL); break;
 				case 5: /*array*/ $set[$method] = print_r($set[$method], true); break;
 				case 9: /*bool*/ $set[$method] = self::_bool_toString($set[$method]); break;
 				default: /*do nothing*/